Fix assembly to use AA prec for filtering
Filtering now works giving a parameter for the minimum AA prec of the score threshold for each tool. That means instead of having a threshold score of 50 for every tool, denovopipeline will select the threshold score for each tool where the AA prec is at least 50, which can translate e.g. to a score of 5 for pNovo 3 and a score of 35 for PointNovo